backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0  # Includes websockets for potential real-time events

# Database (SQLite + ORM)
sqlalchemy==2.0.23
alembic==1.13.1  # For DB migrations (optional; run `alembic init alembic` if needed)

# Data Validation/Models
pydantic==2.5.0
pydantic-settings==2.1.0  # For env var handling (if using .env)

# Authentication & Security
pyjwt==2.8.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
bcrypt==4.0.1
cryptography==41.0.7  # Fernet encryption (fixed key issue)

# File Uploads & Processing
python-multipart==0.0.6  # For FormData in /enroll endpoint
pillow==10.1.0  # Image handling (resize/validate uploads)

# SMS Notifications (Twilio)
twilio==8.10.0  # For sending SMS alerts (e.g., on enrollment/events)

# Push Notifications & Firebase (Admin SDK)
firebase-admin==6.2.0  # For FCM push notifications, Firestore, auth verification

# Development/Testing (Optional)
httpx==0.25.2  # For API testing (e.g., in scripts)
python-dotenv==1.0.0  # Load .env files (for secrets like JWT_KEY, TWILIO_SID, FIREBASE_KEY)
face-recognition==1.3.0  # For easy face encoding/matching
opencv-python==4.8.1.78  # Image processing
numpy==1.26.4 # Arrays for encodings
onnxruntime==1.16.3  # If using ONNX models (optional, lighter than full ML)
python-jose[cryptography]==3.3.0
cmake  # For building some dependencies (like face-recognition)
dlib  # For face recognition (ensure CMake is installed)
flask 
insightface 
requests==2.31.0  # For HTTP calls to ML
scikit-learn==1.3.0  # For cosine_similarity
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
